Remove legacy code
Summary:
Before Scene over Link was supported, we had the idea of a "fake room box" so we could iterate and develop on PC.

Now that Scene over Link is available, this old code isn't used and causes confusion. Removing.

On VirtualRoom.prefab, we just delete some child transforms that were used as fake corner points for the fake room.

Differential Revision: D38823792

fbshipit-source-id: 8ffe5b73f8621b84a03e6eacf8431d9a958139cc