Add mini-samples
Summary:
To ease developer understanding (and celebrate the release of Quest Pro), this adds mini-samples to The World Beyond.

**Problem:** the core Passthrough/Scene SDK samples are intentionally simplistic, but The World Beyond game scene can be too complex.
**Solution:** these samples sit in between the two; simple to understand, but using polished assets to be more inspirational and moddable.

{F785794562}

**PassthroughPet**
This sample demonstrates how to use Unity's Navigation components with Scene API. Moving the blue target around the floor and pressing trigger commands Oppy to walk there.

{F789743610}
* NavMesh created from Scene objects
* Pet walking via NavMeshAgent.SetDestination()
* Animation controller to demonstrate switching between idle/running
* Occlusions using Scene objects
* Control character look direction with SamplePet.DoLookAtBehavior()
* Blob shadow to help visually ground virtual content

**PassthroughRoom**
This sample loads the Scene elements as Passthrough objects, and then destroys the walls and ceiling. Developers could use this as a base for making a wave shooter: replace the virtual environment, and have enemies advance on you from outside into your room.

{F789743665}
* Floor outline as mesh using OVRScenePlaneMeshFilter
* Debris distribution using floor outline and exterior space
* Virtual objects culled when within room
* OVRSceneAndPlaneSwitcher to automatically convert plane anchors (Deck, Couch) to volumes

**VirtualFrames**
Achieving this type of portal effect is possible with different methods; two methods either involve multiple cameras and render buffers (expensive), or stencil buffer tricks (custom shaders for everything). The example shown here is the easiest to set up: all this requires is to use specific prefab overrides in OVRSceneManager.

{F796036336}
* Furniture resizer to demonstrate consistent frame width
* Floor fade material at door base
* Debris at door perimeter
* User notification if missing Scene items

**VoiceTranscription**
To demo some Voice SDK basics, this sample gets and transcribes the user's voice to text. It also includes a button supported by Interaction SDK, as well as a data dialog sample for developers to provide more information to users.

{F789743917}
* Get text transcriptions from the Voice SDK
* Unity Canvas using hands or controllers, powered by Interaction SDK
* SampleNotif prefab, for displaying information and requesting user action
* Sample code for checking and requesting microphone permissions in SampleVoice.CheckPermissionsAndContinue()

NOTE: The app on App Lab remains unchanged; this only impacts the open source project on GitHub

allow-large-files

Reviewed By: dinversini

Differential Revision: D40289707

fbshipit-source-id: 9c40172039b2835aa309c7bbee8e9d4b3730b05c